+++ lokalise/client_methods/teams.py
@@ -3,8 +3,9 @@ lokalise.client_methods.teams
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 This module contains API client definition for teams.
 """
-from typing import Optional, Dict
+from typing import Optional, Dict, Union
 from lokalise.collections.teams import TeamsCollection
+from lokalise.models.team import TeamModel
 from .endpoint_provider import EndpointProviderMixin
 
 
@@ -21,3 +22,14 @@ class TeamMethods(EndpointProviderMixin):
         """
         raw_teams = self.get_endpoint("teams").all(params=params)
         return TeamsCollection(raw_teams)
+
+    def team(self, team_id: Union[str, int]) -> TeamModel:
+        """Fetches a single team by ID.
+
+        :param team_id: ID of the team to fetch
+        :type team_id: int or str
+        :return: Team model
+        """
+        raw_project = self.get_endpoint("teams"). \
+            find(parent_id=team_id)
+        return TeamModel(raw_project)